backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
httpx==0.27.0
Pillow==12.1.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
onnxruntime==1.18.0
transformers==4.48.0
facenet-pytorch==2.5.3
librosa==0.10.2
soundfile==0.12.1
aiofiles==23.2.1
python-dotenv==1.0.1
psutil==5.9.8
